FCAI implementation started

+++ src/ScheduleTechnique.java
@@ -2,6 +2,5 @@
 import java.util.*;
 
 public interface ScheduleTechnique {
-
-    public double calculateExecutionTime(List<Process> processes);
+    public int calculateExecutionTime(List<Process> processes);
 }
